test: consolidate grid header/footer row ITs into one view
Move the two unique assertions from the demo GridViewHeaderAndFooterRows
page/IT (text and component group headers + footer) into the dedicated
GridHeaderFooterRowPage/IT, then delete the demo pair. Component group
header/footer are asserted via getHeaderCellContent/getFooterCellContent.
