+++ vaadin-grid-flow-parent/vaadin-grid-flow-integration-tests/src/test/java/com/vaadin/flow/component/grid/it/GridHeaderFooterRowIT.java
@@ -480,6 +480,73 @@ public class GridHeaderFooterRowIT extends AbstractComponentIT {
         return footerCells;
     }
 
+    @Test
+    public void gridWithHeaderAndFooterRows_headerAndFooterAreRenderered() {
+        grid = $(GridElement.class).id("grid-with-header-and-footer-rows");
+        scrollToElement(grid);
+
+        // Joined columns are at prepended row, i.e. index 0
+        assertRenderedHeaderCell(grid.getHeaderCell(1, 0), "Name", true);
+        assertRenderedHeaderCell(grid.getHeaderCell(1, 1), "Age", true);
+        assertRenderedHeaderCell(grid.getHeaderCell(1, 2), "Street", false);
+        assertRenderedHeaderCell(grid.getHeaderCell(1, 3), "Postal Code",
+                false);
+
+        Assert.assertTrue(
+                "The first column group should have 'Basic Information' header text",
+                grid.getHeaderCellContent(0, 0).getText()
+                        .contains("Basic Information"));
+
+        Assert.assertTrue(
+                "The second column group should have 'Address Information' header text",
+                grid.getHeaderCellContent(0, 1).getText()
+                        .contains("Address Information"));
+
+        Assert.assertTrue("There should be a cell with the renderered footer",
+                grid.getFooterCell(0, 0).getText()
+                        .contains("Total: 500 people"));
+    }
+
+    @Test
+    public void gridWithHeaderWithComponentRenderer_headerAndFooterAreRenderered() {
+        grid = $(GridElement.class).id("grid-header-with-components");
+        scrollToElement(grid);
+
+        assertRenderedHeaderCell(grid.getHeaderCell(1, 0), "<span>Name</span>",
+                true);
+        assertRenderedHeaderCell(grid.getHeaderCell(1, 1), "<span>Age</span>",
+                true);
+        assertRenderedHeaderCell(grid.getHeaderCell(1, 2),
+                "<span>Street</span>", false);
+        assertRenderedHeaderCell(grid.getHeaderCell(1, 3),
+                "<span>Postal Code</span>", false);
+
+        Assert.assertTrue(
+                "The first column group should have a 'Basic Information' component header",
+                grid.getHeaderCellContent(0, 0).getDomProperty("innerHTML")
+                        .contains("<span>Basic Information</span>"));
+
+        Assert.assertTrue(
+                "The second column group should have an 'Address Information' component header",
+                grid.getHeaderCellContent(0, 1).getDomProperty("innerHTML")
+                        .contains("<span>Address Information</span>"));
+
+        Assert.assertTrue("The footer should contain the renderered component",
+                grid.getFooterCellContent(0, 0).getDomProperty("innerHTML")
+                        .contains("<span>Total: 500 people</span>"));
+    }
+
+    private void assertRenderedHeaderCell(GridTHTDElement headerCell,
+            String text, boolean withSorter) {
+        String html = headerCell.getInnerHTML();
+        if (withSorter) {
+            Assert.assertTrue(html.contains("<vaadin-grid-sorter"));
+        } else {
+            Assert.assertFalse(html.contains("<vaadin-grid-sorter"));
+        }
+        Assert.assertTrue(html.contains(text));
+    }
+
     private void clickButton(String id) {
         findElement(By.id(id)).click();
     }
